refactor sha1
